Use intptr_t instead of plain int in OsNotification and its successors to allow passing pointers as their arguments.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@9706 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaAdapterLib/sipXmediaMediaProcessing/src/test/CpPhoneMediaInterfaceTest.cpp
@@ -42,7 +42,7 @@ public:
    StoreSignalNotification() {}
    virtual ~StoreSignalNotification() {}
 
-   OsStatus signal(const int eventData) 
+   OsStatus signal(const intptr_t eventData) 
    { 
       UtlInt* pED = new UtlInt(eventData);
       return (mEDataList.insert(pED) == pED) ?
