+++ sipXportLib/src/os/OsEvent.cpp
@@ -35,7 +35,7 @@
 /* ============================ CREATORS ================================== */
 
 // Constructor
-OsEvent::OsEvent(const int userData)
+OsEvent::OsEvent(const intptr_t userData)
 :  mEventData(-1),
    mIsSignaled(FALSE),
    mSignalSem(OsBSem::Q_PRIORITY, OsBSem::EMPTY),
@@ -58,7 +58,7 @@ OsEvent::~OsEvent()
 // Set the event data and signal the occurrence of the event.
 // Return OS_ALREADY_SIGNALED if the event has already been signaled
 // (and has not yet been cleared), otherwise return OS_SUCCESS.
-OsStatus OsEvent::signal(const int eventData)
+OsStatus OsEvent::signal(const intptr_t eventData)
 {
    OsStatus res;
    OsLock lock(mMutex);
@@ -112,7 +112,7 @@ OsStatus OsEvent::wait(const OsTime& rTimeout)
    return mSignalSem.acquire(rTimeout);
 }
 
-OsStatus OsEvent::setUserData(int userData)
+OsStatus OsEvent::setUserData(intptr_t userData)
 {
    mUserData = userData;
    return OS_SUCCESS;
@@ -123,7 +123,7 @@ OsStatus OsEvent::setUserData(int userData)
 // Return the event data that was signaled by the notifier task.
 // Return OS_NOT_SIGNALED if the event has not been signaled (or has
 // already been cleared), otherwise return OS_SUCCESS.
-OsStatus OsEvent::getEventData(int& rEventData)
+OsStatus OsEvent::getEventData(intptr_t& rEventData)
 {
    OsLock lock(mMutex);
 
@@ -140,7 +140,7 @@ OsStatus OsEvent::getEventData(int& rEventData)
 
 // Return the user data specified when this object was constructed.
 // Always returns OS_SUCCESS.
-OsStatus OsEvent::getUserData(int& rUserData) const
+OsStatus OsEvent::getUserData(intptr_t& rUserData) const
 {
    rUserData = mUserData;
    return OS_SUCCESS;
